feat/fix: new project folder structure
